Move subtensor lift rewrites to their own module
