+++ pytensor/tensor/rewriting/__init__.py
@@ -15,4 +15,5 @@ import pytensor.tensor.rewriting.ofg
 import pytensor.tensor.rewriting.shape
 import pytensor.tensor.rewriting.special
 import pytensor.tensor.rewriting.subtensor
+import pytensor.tensor.rewriting.subtensor_lift
 import pytensor.tensor.rewriting.uncanonicalize
